common_spdk: remove additional unneeded libraries
DPDK 18.02 enabled disabling building of the procinfo
application.  Disabling that enables us to disable
the kvargs, timer, metrics and gso libraries as well -
none of which SPDK uses.

On my dev box, reduces a fresh SPDK build (with DPDK
submodule) from 10.205 to 9.022 seconds.  Reduces
an SPDK rebuild (with no changes) from 4.390 to 4.039
seconds.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I7b8eb7e840372d06594ec3b790b920940ecfcd32
Reviewed-on: https://review.gerrithub.io/403077
Reviewed-by: Daniel Verkamp <daniel.verkamp@intel.com>
Tested-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>