memory: implement "single file segments" for legacy mem mode
DPDK legacy memory mode can now allocate a single
non-physically-contiguous hugetlbfs file on the master NUMA node.
Single file segments are already supported by the new memory
allocation mode, but SPDK holds off with using it just yet.

This enables virtio initiator to work with 2MB hugepages. For virtio,
we have to send a descriptor for each mmap region to the vhost
process, so it can remap the memory and perform DMA on it. When
using 2MB hugepages, this implies sharing possibly hundreds of
descriptors per connection, and managing possibly thousands of
them per vhost.

Due to this DPDK behavior, virtio initiator was practically
limited to 1GB hugepages. This patch lifts the limitation.

Change-Id: Ibfedf510001e08e50c52ee7c91ab12668c243650
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>
Reviewed-on: https://review.gerrithub.io/403596
Reviewed-on: https://review.gerrithub.io/419563
Reviewed-on: https://review.gerrithub.io/428004
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Tested-by: Jim Harris <james.r.harris@intel.com>