rte_vhost2: add asynchronous fd_man
fd_man is an util library for polling file descriptors.
It creates a background poll thread and calls provided
callbacks once a descriptor becomes readable or writeable.
To simplify the threading model for users, it is possible
to call custom callbacks on the background thread whenever
requested.

This is the core of the upcoming Vhost-user backend
implementation.

Change-Id: Ic0f5530c6ccf6470c9e40ffbe73e77121a496ffa
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>