memalloc: keep in mind a failed MAP_FIXED mmap may still perform an unmap
This isn't documented in the manuals, but a failed
mmap(..., MAP_FIXED) may still unmap overlapping
regions. In such case, we need to remap these regions
back into our address space to ensure mem contiguity.
We do it unconditionally now on mmap failure just to
be safe.

Verified on Linux 4.9.0-4-amd64. I was getting
ENOMEM when trying to map hugetlbfs with no space
left, and the previous anonymous mapping was still
being removed.

Fixes: 582bed1e1d1d ("mem: support mapping hugepages at runtime")

Change-Id: I78719304663d9ff93e26e2956e7e7fba2c91213e
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>