vhost2: add initial implementation
This is the initial public API. Although it incorporates
some of the Vhost-user specifics, the API is meant to be
as generic is possible. It will hopefully allow completely
new vhost transports/protocols to reuse the same interface
in future. Subsequent patches in this series add extra
functionalities and extend this interface a bit, although
its general structure remains the same.

This patch also comes with a `transport` layer that's meant
to propagate all user requests to the proper transport
specific handlers. This layer makes the rte_vhost2.h API
fully implemented, but no transports are available yet,
hence there's no point in using this whole lib just yet.
vhost-user socket transport is added in subsequent patches.

Meson build is not supported yet.

Change-Id: If33a657fca3d8ac0324016c50a379c64502ba17e
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>