vhost2: add vhost-user helper layer
This helper layer contains most of the vhost-user message
parsing logic. It is, however, transport-agnostic and doesn't
implement some transport-specific messages like SET_MEM_TABLE,
as those are meant to be implemented by the upper layers. This
layer (further called 'vhost') will be used in the vhost-user
transport coming later within this patch series and
virtio-vhost-user coming at some time in the future.

This layer still references file descriptors here and there,
because those are an integral part of the vhost-user protocol.
However, they are not polled or used in any way and every
transport is allowed to provide -1 instead of any real fd.

Change-Id: I411d4a0f6db7d871620986d117444bd0f914a3aa
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>