vhost2: add asynchronous fd_man
fd_man is an util library for polling file descriptors.
It creates a background poll thread and calls provided
callbacks once a descriptor becomes readable or writeable.
To simplify the threading model for users, it is possible
to call custom callbacks on the background thread whenever
requested.

This fdman spawns a background poll thread and uses
an internal pipe for thread synchronization. Each external
API function puts a callbacks into the pipe that will be
later called by the poll thread. This way, we can simplify
the thread-safety rules and get rid of any mutexes.

This is based on the original fd_man from rte_vhost.

Change-Id: I8ce9b2f06f555e99bcd63a3b82d12b7fd59b842b
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>