vhost2/vhost: handle queue_stop/device_destroy failure
If a stopping or destroying callback fails, vhost will try
to fire the same callback again and again.

We cannot just retry firing the callback straight from the
completion function, as we'd be prone to call stack overflow
then. Callbacks that can fail are now fired in a loop until
they either succeed or are deferred.

Change-Id: Ia4132638fd25e0bbe94446d6c21ca822f95df52f
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>