memory: make eal_get_virtual_area() aware of base-virtaddr alignment
Whenever a calculated base-virtaddr offset had to be
manually aligned to requested page_sz, we did not take
account of that alignment in incrementing the base-virtaddr
offset further. The next requested virtual area could print
a warning "hint [...] not respected!" and let the system
pick an address instead. As a result, this breaks secondary
process support on many system configurations.

Fixes: b7cc54187ea4 ("mem: move virtual area function in common directory")

Change-Id: I93ddcc91d6667543174dfe87a4a95c593ffc58ea
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>