Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> what about common_spdk? The T=spdk-linuxapp-gcc target seems to be left behind now.

I don't know what you mean with that last comment about the T being left behind.  Some of the changes for crypto for DPDK build are already in, what we did was to comment out the options that we want to specify in the makefile dynamically. Turns out this one can't be specified dynamically like the others (see mast dpdkbuild/Makefile) so I needed to add it back in, that's all.  This is how it was before.

Patch-set: 1
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>