bus/pci: fix unexpected resource mapping override
When scanning an already plugged device, the virtual address
of mapped PCI resource in rte_pci_device will be overridden
with 0, that may cause driver does not work correctly.
The fix is not to update any rte_pci_device's field if the being
scanned device's driver is already probed.

Bugzilla ID: 85
Fixes: c752998b5e2e ("pci: introduce library and driver")
Cc: stable@dpdk.org

Change-Id: I7b81b4791287dd71072c26597615bffafdccdb5b
Reported-by: Lv Geoffrey <geoffrey.lv@gmail.com>
Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>
Reviewed-on: https://review.gerrithub.io/430501
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: Jim Harris <james.r.harris@intel.com>