vdev: free memory allocated when sending sync msg
This is a test fix for a possible bug in DPDK where memory
allocated when scanning the vdev bus in a secondary process
is not freed.

Not yet confirmed that this is a bug nor that it will fix an
issue seen with crypto patch in CI that only occurs with crypto
because the vdev bus doesn't exist without the crypto modules
compiled in.

Change-Id: I11360d0ada4b3a570ed937fd8a6d8bf25ebc9e8f
Signed-off-by: paul luse <paul.e.luse@intel.com>
Reviewed-on: https://review.gerrithub.io/425051
Reviewed-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>
Reviewed-on: https://review.gerrithub.io/430059
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Tested-by: Jim Harris <james.r.harris@intel.com>