dev: fix rte_mp_request_sync memleak with multiple requests
When sending multiple requests, rte_mp_request_sync
can succeed sending a few of those requests, but then
fail on a later one and in the end return with rc=-1.
The upper layer - hotplug - currently handles this
case as if no messages were sent and no memory for
response buffers was allocated, which is wrong.

Fixing the above is not so obvious, as rte_mp_request_sync
can fail before the response buffer pointer is even
set (it is unitialized by default). So the caller
cannot safely access it (or free it) just if the
function returns with rc=-1. One way to fix this is
to always initialize the response buffer pointer and
always require the caller to free the response buffer.
This, however, asks for a redesign and should be
addressed with slightly more effort put into it.
Maybe the response buffer pointer should be initialized
by the caller from the very beginning?

This patch makes us go with the ugly solution mentioned
above, so at least our ASAN is silenced.

Change-Id: Iacca2e3f0f8ac9f520f6cf04ca31346258995f75
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>
Reviewed-on: https://review.gerrithub.io/c/440376 (spdk-18.11)