vfio: map all PCI BARs in 2MB-aligned virtual addresses
This is a workaround to a restriction of SPDK env_dpdk library. The
vtophys map in SPDK requires the registered addresses to be 2MB aligned.
This patch enables registering a mapped PCI BAR to the vtophys map of
SPDK by mapping all PCI BARs of all PCI devices to 2MB-aligned virtual
addresses. This is useful in case of peer-to-peer DMA where a PCI device
performs DMA from a PCI BAR of another PCI device. This is the case of
the SPDK vhost target with virtio-vhost-user transport. In specific, the
vhost storage backends perform DMA from a PCI BAR of the
virtio-vhost-user device.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Change-Id: I91061828ac0e38000d5b3ca1e62377e6599440d5