malloc_mp.c: Call memalloc_mem_event_notify from handle_alloc_request
This is necessary to support registering dynamically allocated memory
from secondary processes with the IOMMU. The reason this is necessary is
because the function that registers memory with the IOMMU
(vfio_mem_event_callback) is registered as a mem_event_callback_entry.

RFC because I don't think that I am properly handling the DPDK IPC
aspect of this problem. See try_expand_heap_primary for what I mean
there.

Change-Id: I60de33913f58bc2454069c3844826c92cb043fff
Signed-off-by: Seth Howell <seth.howell@intel.com>