config: always build compressdev and cryptodev API
There's an issue with DPDK and build order when we use makefile
cmd line options like we do to build the PMDs and APIs for
crypto and compression.  Basically DPDK will not complie if we
specify all of them on the command line.  Turns out if we specify
just the APIs in the command line and everything else in the
makefile we can build both crypto and compress just fine. This
allows us to avoid making folks install things like isa-l and
ipsec unless they actually are going to use compress or crypto.

Everyone, hwoever, will build the 2 API modules for DPDK all of
the time but this should not be an issue.

Once this lands, master can be cleaned up to stop specifying
the crypto API on the cmd line and the WIP patch for compress
can do the same.  Until then they are harmless, the config
file will override any options set on the command line.

Change-Id: Iaf2284e646ad2f7b243fa4309b2ef9355591af0b
Signed-off-by: paul luse <paul.e.luse@intel.com>