Update patch set 1
Patch Set 1:

VFIO, for instance, will fail any attempt to map the resources in secondary process if they weren't mapped in primary (I believe it used to crash on the upstream DPDK at the time you created this patch, but that's how it works now. There's an explicit check for that.). I'm not sure if this patch introduces any bugs (thread safety maybe?), but it's not the way to go. If we want to use crypto devices in secondary process, we should first initialize those in the primary process.

Patch-set: 1
Reviewer: Gerrit User 1011218 <1011218@d5d70762-12d0-45a1-890d-524b12d3f735>