vhost: add virtio pci framework
The virtio-vhost-user transport will involve a virtio pci device driver.
There is currently no librte_virtio API that we can use.

This commit is a hack that duplicates the virtio pci code from
drivers/net/.  A better solution would be to extract the code cleanly
from drivers/net/ and share it.  Or perhaps we could backport SPDK's
lib/virtio.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Change-Id: I871f01cfe9116f6bd102bcfdb3c8f7ceff55abd9