vhost: add virtio-vhost-user transport
This patch introduces the virtio-vhost-user transport. This transport is
based on the virtio-vhost-user device. This device replaces the AF_UNIX
socket used by the vhost-user protocol with a virtio device that tunnels
vhost-user protocol messages.  This allows a guest to act as a vhost
device backend for other guests.

For more information on virtio-vhost-user, see
https://wiki.qemu.org/Features/VirtioVhostUser.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Change-Id: I796792f9b620f56000e1b1bdbd35336844b5ad38