vhost: add support for vvu's virtio PCI capabilities
Enhance the vvu device driver so that it utilizes the device's
additional resource capabilities. In specific, this patch adds support
for the doorbells and shared_memory capabilities. The former is used to
find the location of the device doorbells. The latter is used to find
the location of the vhost memory regions in the device's memory address
space. Also, support has been added for the notification capability,
though this configuration structure is not currently being used by the
vvu driver due to DPDK's poll-mode nature.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Change-Id: I95c26f14f4328fda06946bdded644573ebef3d30