[RFC] demonstrate PCIe error notification via VFIO
This patch demonstrates how to get PCIe error
notifications via VFIO for an attached device.  This
has been tested with the aer-inject utility and
a kernel with CONFIG_PCIEAER_INJECT=y.

Note that this only notifies on uncorrectable errors.
It does not specify exactly what error occurred, but
that could be gleaned via PCI configuration space
parsing.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I530ac9de521716f07fc6fec08c6f6f2f33abd213