timer: remove check_tsc_flags()
This code was added 7+ years ago:

commit fb022b85bae4 ("timer: check TSC reliability")

presumably when variant TSCs were still somewhat
common?  But this code doesn't do anything except print
a warning, and the warning doesn't give any kind of
advice to the user, so let's just remove it.

While the warning has no functional meaning, the
/proc/cpuinfo parsing consumes a non-trivial amount
of time which is especially noticeable in secondary
processes.  On my test system, it consumes
21ms out of the 66ms total execution time for
rte_eal_init() in a secondary process.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ie2236f83094672548327dba945b33e3f28fee337