build: remove rte_ethdev from pci bus dependencies
the librte_pci_bus driver lists librte_ethdev as a dependency, but none
of the files in that library depend on the rte_ethdev headers, so I
can't see how this could be a valid dependency. It also breaks our
shared build on freebsd because rte_ethdev is disabled on that machine
by default, so when we try to build the shared library, we get caught up
on this line:

O_TO_S = $(LD) -L$(RTE_SDK_BIN)/lib $(_CPU_LDFLAGS) $(EXTRA_LDFLAGS) \
          -shared $(OBJS-y) $(NO_UNDEFINED) $(LDLIBS) -Wl,-soname,$(LIB)
          -o $(LIB)

which forces all of the dependencies for a given library to exist when
building and linking it.

Change-Id: Ice1305861072e588ae17a71603142702d31b86d4
Reviewed-on: https://review.gerrithub.io/c/spdk/dpdk/+/463940 (spdk-19.05)
Signed-off-by: Seth Howell <seth.howell@intel.com>
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.gerrithub.io/c/spdk/dpdk/+/471067
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>