vhost: fix vring requests validation broken if no FD
[ upstream commit 1407b0752eee2353160701d0182c7e911ccb538a]

When VHOST_USER_VRING_NOFD_MASK is set, the fd_num is 0,
so validate_msg_fds() will return error. In this case,
the negotiation of vring message between vhost user front end and
back end would fail, and as a result, vhost user link could NOT be up.

How to reproduce:
1.Run dpdk testpmd insides VM, which locates at host with ovs+dpdk.
2.Notice that inside ovs there are endless logs regarding failure to
handle VHOST_USER_SET_VRING_CALL, and link of vm could NOT be up.

Fixes: bf472259dde6 ("vhost: fix possible denial of service by leaking FDs")

Signed-off-by: Zhike Wang <wangzk320@163.com> (dpdk-stable)
Reviewed-by: Maxime Coquelin <maxime.coquelin@redhat.com> (dpdk-stable)
Reviewed-by: Tiwei Bie <tiwei.bie@intel.com> (dpdk-stable)
Change-Id: I9e323dc5c5fbfc40eabdf58a9c466229609642da
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.gerrithub.io/c/spdk/dpdk/+/475130
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Tested-by: Jim Harris <james.r.harris@intel.com>