net/netvsc: do RSS across Rx queue only
[ upstream commit 019a28912bf0113566ed5599192973bc1a7cbb07 ]

If number of tx queues is greater than the number of rx queues;
the driver ends up allocating more channels than rx queues.
The problem is that the RSS indirection table is programmed such
that some packets will end up on a channel that would never be
polled. The fix is to limit the RSS indirection table by number
of rx queues not channels.

Fixes: 92d23a57cafe ("net/netvsc: support configuring RSS parameters")

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>