net/mlx5: fix flow META item validation
[ upstream commit 5f3541724e0812e23532fa7e9e89254b3e0c87c6 ]

When flow is inserted with meta match item it requires a certain
register support.
As part of the flow validation of such flows, the validation
function is missing a check that the mlx5 driver is not in
legacy mode in terms of extended meta data support
(MLX5_XMETA_MODE_LEGACY flag).
If the driver is in legacy mode it will cause downstream
function that allocates needed register for meta data.

The fix checks explicitly the conditions for support of
meta data in FDB mode. If the conditions are not met
an error message will be issued.

Fixes: 9bf26e1318e3 ("ethdev: move egress metadata to dynamic field")

Signed-off-by: Shy Shyman <shys@mellanox.com>
Acked-by: Matan Azrad <matan@mellanox.com>