examples: fix flattening directory layout on install
[ upstream commit e5b95003f1df2ac65cbd11e642b1d6a91bd1058a ]

By installing the examples one-by-one in a loop in the examples
meson.build file we effectively flattened out the structure of the examples
folder and omitted some common and shared subfolders that were never
directly built.  Instead, we can remove the loop and just have the whole
"examples" folder installed as-is in a single statement, preserving its
directory structure, and thereby fixing the build of a number of the
examples.

Fixes: 2daf565f91b5 ("examples: install as part of ninja install")

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>