meson: remove checks for optional libraries
Very few libraries in DPDK are marked as optional.
For SPDK when most of the drivers are disabled,
the requirements are much lower.

By removing the check for optional libraries,
it is possible to pass a narrow set of actually required
libraries.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/dpdk/+/15920 (spdk-22.11.1)

(cherry picked from commit eb2f7341a040f888a7ac6df10238e894f863be00)
Change-Id: I60fbc7307a4f33482025a3b3c00948c091d236ff
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Signed-off-by: Krzysztof Sprzaczkowski <krzysztof.sprzaczkowski@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/dpdk/+/18464
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>