mem: allow using ASan in multi-process mode
Multi-process applications operate on shared hugepage memory but each
process has its own ASan shadow region which is not synchronized with
the other processes. This causes issues when different processes try to
use the same memory because they have their own view of which addresses
are valid.

Fix it by mapping the shadow regions for memseg lists as shared memory.
The primary process is responsible for creating and removing the shared
memory objects.

Disable ASan instrumentation for triggering the page fault in
alloc_seg() because if the segment is already allocated by another
process and is marked as free in the shadow, accessing this address will
cause an ASan error.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-on: https://review.spdk.io/c/spdk/dpdk/+/26629 (spdk-25.07)
Reviewed-on: https://review.spdk.io/c/spdk/dpdk/+/27396 (spdk-25.11)

(cherry picked from commit e38aa563d651b46e0f45037cbba77cdc42f0b84f)
Change-Id: I2bb6ae100d080aad30ee44d7e5d200962f74d1a8
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/dpdk/+/28659
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>