bus/pci: don't open uio device in secondary process
The uio_pci_generic driver clears the bus master bit when the device
file is closed. So, when the secondary process terminates after probing
a device, that device becomes unusable in the primary process.

To avoid that, the device file is now opened only in the primary process
and the secondary gets it over UNIX domain socket via SCM_RIGHTS.

Fixes: 847d78fb9530 ("bus/pci: fix FD in secondary process")
Cc: stable@dpdk.org

Following patch was not merged to DPDK upstream, nor spdk-24.07.
Reviewed-on: https://review.spdk.io/c/spdk/dpdk/+/25951 (spdk-25.03)
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>

Change-Id: I7bcb9d682f42ad5796e17ce8fc20a4a88eea5921
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/dpdk/+/26634
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>